requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
Pillow>=9.0
